Update README.md (mod support PSP patch error)
so people don't spend an ungodly amount of time trying to get it to work like i did. The patches do work, but it's specifically the mod support one that breaks the game at launch with a black screen, and there's not a real way to refresh it/apply the mod patch mid session.